+++ src/test/java/kr/co/tmax/arenademobe/service/CompanyStockServiceTest.java
@@ -1,18 +1,13 @@
 package kr.co.tmax.arenademobe.service;
 
 import kr.co.tmax.arenademobe.domain.CompanyStock;
-import kr.co.tmax.arenademobe.domain.Member;
 import kr.co.tmax.arenademobe.repository.CompanyStockRepository;
-import kr.co.tmax.arenademobe.repository.MemberRepository;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import kr.co.tmax.arenademobe.repository.UserAccountRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class CompanyStockServiceTest {
@@ -21,15 +16,15 @@ class CompanyStockServiceTest {
     private CompanyStockRepository companyStockRepository;
 
     @Autowired
-    private MemberRepository memberRepository;
+    private UserAccountRepository memberRepository;
 
     @Autowired
     private CompanyStockService companyStockService;
 
-    @BeforeEach
-    public void 회원미리저장() {
-        memberRepository.save(new Member(1L, "userA", "삼성전자", "LG에너지솔루션", "현대차"));
-    }
+//    @BeforeEach
+//    public void 회원미리저장() {
+//        memberRepository.save(new Member(1L, "userA", "삼성전자", "LG에너지솔루션", "현대차"));
+//    }
 
     @Test
     public void 추천로직테스트() {
